+++ dev/youngmin/3_class/Docker/requirements.txt
@@ -1,0 +1,8 @@
+librosa==0.10.0
+numpy>=1.23
+matplotlib
+tqdm
+scikit-learn
+soundfile
+tensorflow==2.13.0
+seaborn
