+++ dev/youngmin/pytorch/scripts/train.py
@@ -1,0 +1,101 @@
+import numpy as np
+import torch
+import torch.nn as nn
+import torch.optim as optim
+from torch.utils.data import DataLoader, TensorDataset
+from sklearn.model_selection import train_test_split
+from pathlib import Path
+
+# 1. 모델 정의
+class AudioCNN(nn.Module):
+    def __init__(self):
+        super(AudioCNN, self).__init__()
+        self.conv1 = nn.Conv2d(1, 32, kernel_size=3, padding=1)
+        self.pool1 = nn.MaxPool2d(2, 2)
+        self.conv2 = nn.Conv2d(32, 64, kernel_size=3, padding=1)
+        self.pool2 = nn.MaxPool2d(2, 2)
+        self.fc1 = nn.Linear(4032, 64)  # 여기 수정!
+        self.dropout = nn.Dropout(0.3)
+        self.fc2 = nn.Linear(64, 1)
+
+    def forward(self, x):
+        x = torch.relu(self.pool1(self.conv1(x)))
+        x = torch.relu(self.pool2(self.conv2(x)))
+        x = x.view(x.size(0), -1)
+        x = torch.relu(self.fc1(x))
+        x = self.dropout(x)
+        x = torch.sigmoid(self.fc2(x))
+        return x
+
+
+# 2. 데이터 불러오기
+x = np.load("../data/x.npy")  # shape: (9374, 86, 14)
+y = np.load("../data/y.npy")  # shape: (9374,)
+
+# 🔸 float32 변환 (PyTorch는 float32 권장)
+x = x.astype(np.float32)
+y = y.astype(np.float32)
+
+# 🔸 채널 차원 추가 (PyTorch용: (N, 1, H, W))
+x = np.expand_dims(x, axis=1)  # shape: (9374, 1, 86, 14)
+
+# 🔸 표준 정규화 (z-score)
+x = (x - x.mean()) / x.std()
+
+# 🔸 학습/검증 분리
+x_train, x_val, y_train, y_val = train_test_split(x, y, test_size=0.2, random_state=42)
+
+# 🔸 텐서로 변환
+train_dataset = TensorDataset(torch.tensor(x_train), torch.tensor(y_train))
+val_dataset = TensorDataset(torch.tensor(x_val), torch.tensor(y_val))
+
+train_loader = DataLoader(train_dataset, batch_size=32, shuffle=True)
+val_loader = DataLoader(val_dataset, batch_size=32)
+
+
+# 3. 학습 준비
+device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
+model = AudioCNN().to(device)
+
+criterion = nn.BCELoss()
+optimizer = optim.Adam(model.parameters(), lr=0.001)
+
+# 4. 학습 루프
+best_val_loss = float('inf')
+patience, wait = 5, 0
+
+for epoch in range(30):
+    model.train()
+    running_loss = 0.0
+    for xb, yb in train_loader:
+        xb, yb = xb.to(device), yb.to(device).unsqueeze(1)
+        optimizer.zero_grad()
+        preds = model(xb)
+        loss = criterion(preds, yb)
+        loss.backward()
+        optimizer.step()
+        running_loss += loss.item()
+
+    model.eval()
+    val_loss = 0.0
+    with torch.no_grad():
+        for xb, yb in val_loader:
+            xb, yb = xb.to(device), yb.to(device).unsqueeze(1)
+            preds = model(xb)
+            val_loss += criterion(preds, yb).item()
+
+    avg_train_loss = running_loss / len(train_loader)
+    avg_val_loss = val_loss / len(val_loader)
+
+    print(f"[Epoch {epoch+1}] Train Loss: {avg_train_loss:.4f} | Val Loss: {avg_val_loss:.4f}")
+
+    if avg_val_loss < best_val_loss:
+        best_val_loss = avg_val_loss
+        wait = 0
+        torch.save(model.state_dict(), "../model/cnn_model.pt")
+        print("  📌 모델 저장됨")
+    else:
+        wait += 1
+        if wait >= patience:
+            print("  🛑 Early Stopping")
+            break
